Fixed handling of elements from foreign namespaces in values object
When encountering an element from another namespace the value object parser
does skip the opening element. However when it encounters the closing element
it did handle it like it was the closing element of the one being processed.

This commit and unit test fixes the issue by ignoring the element completely